KAS-3107 Edit proof within received panel
