--- app/components/publications/publication/proofs/proof-edit-modal.hbs
@@ -1,38 +1,0 @@
-<Auk::Modal>
-  <Auk::Modal::Header @title={{t "edit-proof"}} @onClose={{@onCancel}} />
-  <Auk::Modal::Body>
-    <div class="auk-form-group">
-      <Auk::Label for="proofUploadReceivedDate" @required={{true}}>
-        {{t "date"}}
-      </Auk::Label>
-      <Auk::Datepicker
-        id="proofUploadReceivedDate"
-        @date={{this.receivedDate}}
-        @onChange={{this.setReceivedDate}}
-      />
-    </div>
-    <div class="auk-form-group">
-      <Auk::Label for="proofPrintCorrector">
-        {{t "who-proof-reads"}}
-      </Auk::Label>
-      <Auk::Input
-        id="proofPrintCorrector"
-        @value={{mut this.proofPrintCorrector}}
-      />
-    </div>
-  </Auk::Modal::Body>
-  <Auk::Modal::Footer
-    @cancelDisabled={{this.isCancelDisabled}}
-    @onCancel={{@onCancel}}
-  >
-    <Auk::Button
-      data-test-proof-edit-save
-      @skin="primary"
-      @disabled={{this.isSaveDisabled}}
-      @loading={{this.save.isRunning}}
-      {{on "click" (perform this.save)}}
-    >
-      {{t "save"}}
-    </Auk::Button>
-  </Auk::Modal::Footer>
-</Auk::Modal>
