+++ app/components/publications/publication/proofs/proof-received-panel.js
@@ -1,18 +1,43 @@
 import Component from '@glimmer/component';
 import { action } from '@ember/object';
 import { tracked } from '@glimmer/tracking';
+import { task } from 'ember-concurrency-decorators';
+import { isEmpty } from '@ember/utils';
 
 /**
  * @argument {proofingActivity}
  * @argument {publicationSubcase}
- * @argument {onSaveEditProof}
+ * @argument {onSaveEditReceivedProof}
  */
 export default class PublicationsPublicationProofProofReceivedPanel extends Component {
   @tracked showEditProofModal = false;
 
+  @tracked newReceivedDate;
+  @tracked newProofPrintCorrector;
+
+  get isCancelDisabled() {
+    return this.save.isRunning;
+  }
+
+  get isSaveDisabled() {
+    return isEmpty(this.newReceivedDate);
+  }
+
+  @task
+  *save() {
+    yield this.args.onSaveEditReceivedProof({
+      proofingActivity: this.args.proofingActivity,
+      receivedDate: this.newReceivedDate,
+      proofPrintCorrector: this.newProofPrintCorrector,
+    });
+    this.showEditProofModal = false;
+  }
+
   @action
   openEditProofModal() {
     this.showEditProofModal = true;
+    this.newReceivedDate = this.args.proofingActivity.endDate;
+    this.newProofPrintCorrector = this.args.publicationSubcase.proofPrintCorrector;
   }
 
   @action
@@ -21,8 +46,12 @@ export default class PublicationsPublicationProofProofReceivedPanel extends Comp
   }
 
   @action
-  async saveEdit(proofEdit) {
-    await this.args.onSaveEditReceivedProof(proofEdit);
-    this.showEditProofModal = false;
+  setNewReceivedDate(selectedDates) {
+    if (selectedDates.length) {
+      this.newReceivedDate = selectedDates[0];
+    } else {
+      // this case occurs when users manually empty the date input-field
+      this.newReceivedDate = undefined;
+    }
   }
 }
